Move QgsGpsMarker to gui
