+++ src/gui/gps/qgsgpsmarker.cpp
